Move WorkflowDefinition building to dedicated WorkflowDefinitionBuilder
