+++ library/core/engine/src/main/kotlin/de/lise/fluxflow/engine/workflow/WorkflowActivationServiceImpl.kt
@@ -1,18 +1,12 @@
 package de.lise.fluxflow.engine.workflow
 
 import de.lise.fluxflow.api.workflow.Workflow
-import de.lise.fluxflow.api.workflow.WorkflowDefinition
 import de.lise.fluxflow.api.workflow.WorkflowIdentifier
 import de.lise.fluxflow.api.workflow.action.WorkflowAction
 import de.lise.fluxflow.persistence.workflow.WorkflowData
-import de.lise.fluxflow.stereotyped.metadata.MetadataBuilder
-import de.lise.fluxflow.stereotyped.workflow.ModelListenerDefinitionBuilder
-import de.lise.fluxflow.stereotyped.workflow.action.WorkflowActionDefinitionBuilder
 
 class WorkflowActivationServiceImpl(
-    private val modelListenerDefinitionBuilder: ModelListenerDefinitionBuilder,
-    private val metadataBuilder: MetadataBuilder,
-    private val actionDefinitionBuilder: WorkflowActionDefinitionBuilder,
+    private val workflowDefinitionBuilder: WorkflowDefinitionBuilder,
 ) : WorkflowActivationService {
     @Suppress("UNCHECKED_CAST")
     override fun <TModel> activate(data: WorkflowData): Workflow<TModel> {
@@ -22,8 +16,7 @@ class WorkflowActivationServiceImpl(
     @Suppress("UNCHECKED_CAST")
     private fun <TModel : Any> activateTypeSafe(data: WorkflowData, model: TModel): Workflow<TModel> {
         val actions = mutableListOf<WorkflowAction<TModel>>()
-        val definition = buildDefinition(model)
-
+        val definition = workflowDefinitionBuilder.build(model)
         val workflow = WorkflowImpl(
             definition,
             WorkflowIdentifier(data.id),
@@ -39,18 +32,5 @@ class WorkflowActivationServiceImpl(
         )
 
         return workflow
-    }
-
-    private fun <TModel : Any> buildDefinition(model: TModel): WorkflowDefinition<TModel> {
-        val modelListenerDefinitions = modelListenerDefinitionBuilder.build(model)
-        val modelType = model::class
-        val metadata = metadataBuilder.build(modelType)
-        val actionDefinitions = actionDefinitionBuilder.build(modelType)
-        
-        return WorkflowDefinitionImpl(
-            modelListenerDefinitions,
-            metadata,
-            actionDefinitions
-        )
     }
 }
